New Branch

+++ src/Domain/Useful/Position.java
@@ -1,29 +1,29 @@
 package Domain.Useful;
 
 public class Position {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     public Position(){}
 
-    public void setX(Integer x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(Integer y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public Integer getX() {
+    public double getX() {
         return x;
     }
 
-    public Integer getY() {
+    public double getY() {
         return y;
     }
